gen 12

+++ source/d1dun.cpp
@@ -1728,7 +1728,7 @@ void D1Dun::loadObject(int objectIndex)
         }
     }
     const DunObjectStruct *objStr = &DunObjConvTbl[objectIndex];
-    if (i >= this->customObjectTypes.size() && (unsigned)objectIndex < lengthof(DunObjConvTbl) && objStr->type != 0 && !this->assetPath.isEmpty()) {
+    if (i >= this->customObjectTypes.size() && (unsigned)objectIndex < (unsigned)lengthof(DunObjConvTbl) && objStr->type != 0 && !this->assetPath.isEmpty()) {
         int objFileIndex = objStr->animType;
         result.frameNum = objStr->frameNum;
         QString celFilePath = this->assetPath + "/Objects/" + objfiledata[objFileIndex].path + ".CEL";
@@ -1798,7 +1798,7 @@ void D1Dun::loadMonster(int monsterIndex)
         }
     }
     const DunMonsterStruct *monStr = &DunMonstConvTbl[monsterIndex];
-    if (i >= this->customObjectTypes.size() && (unsigned)monsterIndex < lengthof(DunMonstConvTbl) && monStr->type != 0 && !this->assetPath.isEmpty()) {
+    if (i >= this->customObjectTypes.size() && (unsigned)monsterIndex < (unsigned)lengthof(DunMonstConvTbl) && monStr->type != 0 && !this->assetPath.isEmpty()) {
         int moFileIndex = monStr->animType;
         QString cl2FilePath = this->assetPath + "/Monsters/" + monfiledata[moFileIndex].path + "N.CL2";
         QString trnFilePath;
